+++ src/main/java/com/ing/data/cassandra/jdbc/CassandraPreparedStatement.java
@@ -248,7 +248,7 @@ public class CassandraPreparedStatement extends CassandraStatement
         doExecute();
         // Return true if the first result is a non-null ResultSet object; false if the first result is an update count
         // or there is no result.
-        return this.currentResultSet != null && ((CassandraResultSet) this.currentResultSet).isQuery();
+        return this.currentResultSet != null && ((CassandraResultSet) this.currentResultSet).isQueryResult();
     }
 
     @SuppressWarnings("ResultOfMethodCallIgnored")